// src/v1/modules/transaction-history/dto/create-transaction-history.dto.ts
import { TokenEnum } from '@/v1/config/settings';
import {
  TransactionDirectionEnum,
  PaymentStatus,
  TransactionTypeEnum,
  YCRampPaymentEventEnum,
} from '@/v1/models/payment.types';
import {
  FeeLevel,
  WalletWebhookEventEnum,
} from '@/v1/models/wallet-manager.types';
import { UserEntity } from '@/v1/utils/typeorm/entities/user.entity';

export class TransactionHistoryDto {
  event?: WalletWebhookEventEnum | YCRampPaymentEventEnum;
  transactionId: string;
  transactionDirection: TransactionDirectionEnum;
  transactionType: TransactionTypeEnum;
  assetCode: TokenEnum;
  amount: string;
  fee?: string;
  feeLevel?: FeeLevel;
  blockchainTxId?: string;
  reason?: string | null;
  createdAt?: Date;
  updatedAt?: Date;
  walletId: string;
  walletName?: string;
  paymentStatus?: PaymentStatus;
  sourceAddress?: string;
  destinationAddress: string;
  paymentNetwork: string;
  tokenId?: string;
  user: UserEntity;
  rampID?: string;
  mainAssetAmount?: number;
  transactionMessage?: string;
}